increase alt hitbox to fill titlebar
This introduces the functionality to perform actions such as closing a maximized window by pressing in the topright-most corner of the screen. This was previously unavailable because the buttons for alt were smaller. The buttons were deliberately kept small without a larger invisible hitbox because at the time, scaling of the application's icon was not well regulated and could easily be warped. This may still be the case and testing did not identify it.